+++ lib/pages/admin_workout_add_screen.dart
@@ -207,16 +207,8 @@ class _AdminWorkoutAddScreenState extends State<AdminWorkoutAddScreen> {
       child: ElevatedButtonWidget(
         widget.workout != null ? 'Save Workout' : 'Add Workout',
         onPressed: () async {
-          if (!nameBloc.isValid()) {
-            return ToastController().show(
-              nameBloc.errorText,
-            );
-          }
-          if (!descriptionBloc.isValid()) {
-            return ToastController().show(
-              descriptionBloc.errorText ?? 'Invalid description',
-            );
-          }
+          if (!nameBloc.isValid()) return;
+          if (!descriptionBloc.isValid()) return;
           try {
             exercisesSel.sort(
               (a, b) =>